Removes extraneous <?php and ?> tags.
This commit removes the extraneous <?php and ?> tags. These are not only removed for code readability, as the PHP code is much easier to read in this new form, but also as with PHP, extraneous spacing and line breaks between the <?php and ?> tags will cause extra spacing and line breaks sent to the client. This also means that there will be saved bandwidth.