Remove referrer meta tag from admin head
It's not clear why this tag was added, and it's even less clear what
benefit there would be to keeping it. Its only real function seems to be
leaking YOURLS' admin panel location (see #2432), which is not desirable.

A hook (or hooks) will likely be added to let plugins control referrer
behavior, by adding the relevant HTTP header(s) and/or meta tag(s).